place consumer under integration test
