--- internal/workqueue/consumer_test.go
@@ -1,1 +1,0 @@
-package workqueue_test
